Add basic Copyright (c) one-line to all files
This does not count towards the 48 line limit.

The opinions of the individual authors may vary as to whether they want
to claim copyright or not, but the rules of the competion allow the
organisers to publish the files as they see fit and it's easier to
relax permission later than lock it down.